+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added support for `Series.str.ljust` and `Series.str.rjust`.
 - Added support for `Series.str.center`.
 - Added support for `Series.str.pad`.
+- Added support for `DataFrame.from_dict` and `DataFrame.from_records`.
 
 
 ## 1.26.0 (2024-12-05)